MOSIP-26582 added BioHashValidator for validating hash content
